[feature/issue-42] router샘플
