+++ package.json
@@ -20,7 +20,7 @@
     "react-dom": "^17.0.2",
     "react-query": "^3.34.19",
     "react-redux": "^7.2.8",
-    "react-router-dom": "^6.2.2",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.0",
     "redux": "^4.1.2",
     "typescript": "^4.4.2",
