+++ src/layouts/App.tsx
@@ -1,3 +1,5 @@
+import Settings from '../pages/Settings';
+import Search from '../pages/Search';
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import Main from '../pages/Main';
@@ -8,6 +10,8 @@ function App() {
   return (
     <Routes>
       <Route path="/" element={<Main />} />
+      <Route path="/search" element={<Search />} />
+      <Route path="/settings" element={<Settings />} />
     </Routes>
   );
 }
