+++ src/pages/Settings.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Settings = () => {
+  return (
+    <>
+      <div>settings</div>
+    </>
+  );
+};
+
+export default Settings;
